docs(roadmap): the P29.5 ledger — what landed, what it measured, and what it carries (P29.5)
The §13 STATUS block for the authoring wave, plus a status note on the P29.5
clause list so it agrees with what shipped.

Carried in it: the seven-AnimModifiers table read off the stock sample's own
uasset name tables (three derived, three superseded, one honest CANNOT on a
channel name inside a binary this reading did not decode); the derivation's
second opinion against the generator's own stride x cadence (4.1% on the walk,
3.9% on the run) and the 2.1x over-read the first estimator produced, which is a
fact about pendulum-legged content rather than about the measure; the mantle arc
against the ease it replaced (0.6 m vs 0.35 m a quarter of the way up a 1 m
ledge, both landing to the millimetre); the preview warm path (56.95 -> 17.59 ms
on an 18k-triangle model, 13.9 -> 3.2 us on the mannequin); the six mutations and
what each killed; and the remainders, including the blend mode and overlay state
deliberately unexposed for a third wave with their owed parity arms named.

Also: clippy's reading of this wave's own code — a complex return type given a
name, and a redundant f64 cast.

+++ crates/inf-anim/src/derive.rs
@@ -1134,7 +1134,7 @@ mod tests {
             &mut clip,
             0,
             vec![0.0, 1.0],
-            vec![q(0.0), q(std::f64::consts::FRAC_PI_4 as f64)],
+            vec![q(0.0), q(std::f64::consts::FRAC_PI_4)],
         );
         clip.duration = 1.0;
